backend/requirements.txt
flask
flask-cors
requests
pypdf
img2pdf
Spire.Presentation
beautifulsoup4
Spire.Doc
